refactor: replace EditorControlView with individual widgets for better modularity
